Do not cause compilation to fail if libapp is not cloned to a directory called `libapp'.
This is done by creating a subdirectory, libapp, within the libapp repo. This is the cleanest solution I can think of because it lets the header files still refer to eachother as libapp/<headername>.h during compiletime as this is required after installation (d50ee37986b).

This will in the future support a release of libapp, where the directory libapp is tarballed into would be called `libapp-0.1'.